==> notebooks/fashion_data/Step1_format_data.ipynb <==
As the first step of this tutorial, we will format our data.

For our dataset, we will be using a dataset from Kaggle: https://www.kaggle.com/paramaggarwal/fashion-product-images-small

Note: You will need to create an account with Kaggle in order to download the data. Once you have done that, download the data to wherever you are going to train your model.

%load_ext autoreload

%autoreload 2

import os
import sys
sys.path.append('../../')

import pandas as pd
from sklearn.model_selection import train_test_split

Read in data
===

Note: change this to wherever you stored your data

DATA_DIR = '/home/ec2-user/fashion_dataset'

df = pd.read_csv(f'{DATA_DIR}/styles.csv', error_bad_lines=False)

Some of the data seems to be corrupt so we will just skip those bad lines for now. 

df.shape

df.head()

Each line in the csv represents a single product, and we have information about who the product is for, what type of product it is, and some attributes of the products, all as separate columns. 

We also have the product name, which will use as the text input to our models.

Each product also has a corresponding image in the `images` folder with the name `{id}.jpg`.

Creating Image URL
===

We need to create a column indicating where the images are stored.

Note: change this to wherever you stored your data

df['image_urls'] = df['id'].apply(lambda x: f'{DATA_DIR}/images/{x}.jpg')

Let's check the first one to make sure it worked.

df.loc[0]['image_urls']

df.head()

Some of the rows do not have matching image files, so we need to remove those.

We'll check if the file exists and if not, then we'll throw out those rows.

bad_ids = []

for i, row in df.iterrows():
    if not os.path.isfile(row['image_urls']):
        print(row['id'])
        print(row['image_urls'])
        bad_ids.append(row['id'])

df.shape

df = df[~df['id'].isin(bad_ids)]

df.shape

We also need to throw out rows without associated text.

df = df[df['productDisplayName'].notnull()]

df.shape

Split into task specific datasets
===

Part of the power of Octopod comes from the fact that you can train one model on different datasets for each task. That way you don't have to label all of your items with the exact same labels.

We will fake having different datasets by subsampling from this main df. For this first version, we'll train on the `gender` and `season` columns.

gender_df = df.sample(frac=0.8)

gender_df.gender.value_counts()

season_df = df.sample(frac=0.6)

season_df.season.value_counts()

Format gender data
===

Now we will put the data in the format needed for Octopod.

gender_df.shape

First, we'll get rid of any null values because Octopod will throw an error if given a null value.

gender_df = gender_df[gender_df['gender'].notnull()]

gender_df.shape

We set the `gender` column to be a pandas category type

gender_df['gender'] = gender_df['gender'].astype('category')

We create a new column called `gender_cat` that contains the category to predict as an integer.
This is an important step because Octopod does not natively handle categories as text.

gender_df['gender_cat'] = gender_df['gender'].cat.codes

# Category mapping
{label: i for i, label in enumerate(gender_df['gender'].cat.categories)}

Note: you will need to save this mapping for later so that you can use your model to make predictions.
Since this is a simple model, we won't save it, but for a real project, we would store it somewhere.

Now our gender data is in the right format for Octopod so we'll split it and save it again.

Note: the columns required for training the models are:
- gender_cat
- image_urls
- productDisplayName
- id (for bookkeeping)

We keep the other columns for convenience, but could drop them if we wanted to.

gender_train_df, gender_valid_df = train_test_split(gender_df, train_size=0.75, random_state=17)

gender_train_df.to_csv(f'{DATA_DIR}/gender_train.csv', index=False)

gender_valid_df.to_csv(f'{DATA_DIR}/gender_valid.csv', index=False)

Format season data
===

season_df.shape

First, we'll get rid of any null values

season_df = season_df[season_df['season'].notnull()]

season_df.shape

We set the `season` column to be a pandas category type

season_df['season'] = season_df['season'].astype('category')

We create a new column called `season_cat` that contains the category to predict as an integer.
This is an important step because Octopod does not natively handle categories as text.

season_df['season_cat'] = season_df['season'].cat.codes

# Category mapping
{label: i for i, label in enumerate(season_df['season'].cat.categories)}

Note: you will need to save this mapping for later so that you can use your model to make predictions.
Since this is a simple model, we won't save it, but for a real project, we would store it somewhere.

Now our season data is in the right format for Octopod so we'll split it and save it again.

season_train_df, season_valid_df = train_test_split(season_df, train_size=0.75, random_state=17)

season_train_df.to_csv(f'{DATA_DIR}/season_train.csv', index=False)

season_valid_df.to_csv(f'{DATA_DIR}/season_valid.csv', index=False)

Now that we have some data, we can train a model with it! Move onto the notebook `Step2_train_image_model.ipynb` to see how.
